split one file structure
