+++ configs/config.yml
@@ -1,5 +1,7 @@
 redis:
-  address: "redis:6379"
+  address:
+  - "redis:6379"
+  - "localhost:6379"
   ttl: "30m"
 image:
   cache:
