+++ docker-compose.yml
@@ -1,11 +1,11 @@
 services:
-  bot:
-    build: .
-    ports:
-      - "8000:8000"
-    depends_on:
-      - redis
-    restart: always
+#  bot:
+#    build: .
+#    ports:
+#      - "8000:8000"
+#    depends_on:
+#      - redis
+#    restart: always
 
   redis:
     image: redis:alpine
