+++ internal/bot/image_cache.go
